requirements.txt
# ==========================================================
# Ultima_RAG - Python Dependencies
# ==========================================================
#
# IMPORTANT: torch, torchvision, torchaudio are intentionally
# NOT in this file. start.bat owns those packages exclusively.
#
# DO NOT add torch/torchvision/torchaudio here under any
# circumstance - pip will install the CPU-only build from
# PyPI and silently break GPU support.
# ==========================================================

# --- ML / AI Core ---
# Note: sentence-transformers and easyocr list torch as a
# dependency. pip may pull in CPU torch when installing them.
# start.bat handles this by force-uninstalling torch AFTER
# requirements.txt runs, then reinstalling the GPU build.
sentence-transformers>=2.2.0,<3.0.0
transformers>=4.45.0
accelerate>=0.26.0
bitsandbytes>=0.43.0
qwen-vl-utils>=0.0.4

# --- Metacognitive Brain (LangGraph) ---
langchain>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.0.10
langchain-core>=0.1.0

# --- SOTA Storage (LanceDB) ---
lancedb>=0.5.0
pyarrow>=14.0.0
tantivy>=0.21.1

# --- Backend & API ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.0.0
requests>=2.28.0

# --- Data Processing ---
numpy==1.26.4
pandas>=2.0.0,<3.0.0
PyPDF2>=3.0.0
nltk>=3.8.0

# --- Multimodal Processing ---
faster-whisper>=1.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
pydub>=0.25.1
easyocr>=1.7.0

# --- Utilities ---
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
huggingface-hub>=0.16.0
tokenizers>=0.13.0
scipy>=1.10.0
fpdf2>=2.7.0

# --- Web Breakout Agent ---
duckduckgo-search>=6.0.0
trafilatura>=1.6.0
